Break out paginator tests
